fix: VCallScreen brand logo
