+++ video-sdk/src/test/java/com/kaleyra/video_sdk/ui/call/screen/VCallScreenLandscapeTest.kt
@@ -7,4 +7,4 @@ import org.robolectric.annotation.Config
 // This class contains the same tests as VCallScreenTest but Robolectric is configured with a landscape device
 @Config(qualifiers = "w360dp-h640dp-land")
 @RunWith(RobolectricTestRunner::class)
-class VCallScreenLandscapeTest: VCallScreenTest()
+internal class VCallScreenLandscapeTest: VCallScreenTest()
